requirements.txt
pylsl
scikit-learn
pyqt5
python-chess
numpy
scipy